checking if double precision works in alx_book_store file

+++ alx_book_store.sql
@@ -11,7 +11,7 @@ CREATE TABLE Book(
   book_id INT PRIMARY KEY,
   title VARCHAR(130),
   author_id  INT FOREIGN KEY REFERENCES author(author_id),
-  price DOUBLE,
+  price DOUBLE PRECISION,
   publication_date DATE
 );
 
@@ -33,7 +33,7 @@ CREATE TABLE Order_Details(
   orderdetailid INT PRIMARY KEY,
   order_id INT,
   book_id INT,
-  quantity DOUBLE,
+  quantity DOUBLE PRECISION,
   FOREIGN KEY (order_id) REFERENCES Orders(order_id),
   FOREIGN KEY (book_id) REFERENCES Books(book_id)
 );
